update:add user expel

+++ src/modules/members/members.service.ts
@@ -25,7 +25,7 @@ export class MembersService {
 
   //convenience function section
   async findMember(boardId: number, userId: number) {
-    const member = await this.MembersRepository.findOne({ where: { userId, boardId } });
+    const member = await this.MembersRepository.findOne({ where: { userId: userId, boardId } });
     return member;
   }
 
@@ -47,13 +47,23 @@ export class MembersService {
   }
 
   async banMember(userIdForBan: number, boardId: number, userId: number) {
-    const member = await this.MembersRepository.findOne({
-      where: { userId: userIdForBan, boardId: boardId },
-    });
+    if (await this.checkMemberRole(boardId, userId, MemberRoles.ONLY_VIEW))
+      throw new NotFoundException({ message: "권한이 없습니다." });
 
+    const target = userIdForBan["userId"];
+
+    const member = await this.MembersRepository.findOne({
+      where: { userId: target, boardId: boardId },
+    });
     if (!member) throw new NotFoundException({ message: "해당 보드에 속한 유저가 아닙니다." });
 
+    if (await this.checkMemberRole(boardId, target, MemberRoles.ADMIN)) {
+      throw new NotFoundException({
+        message: "해당 유저는 당신과 동등하거나 더 높은 권한을 가지고 있습니다.",
+      });
+    }
     this.MembersRepository.delete({ id: member.id });
+    return { message: "해당유저가 성공적으로 강퇴되었습니다." };
   }
 
   async updateMemberRoles(
